requirements.txt
pillow
seaborn # to draw confusion matrix
torch>=1.13.1
torchvision>=0.14.1
tqdm